Add files via upload
This is a console based game. The game has 3 starting rooms so the player can get used to the 3 types of puzzles in the game. The game then will randomize the puzzles and add a random amount of puzzles for the player to solve. Each room should be slightly different from the others.